move the jsx not main to separate comps
