+++ src/component/ProfileTrainer/components/SideBarButton/styles.module.css
@@ -1,0 +1,17 @@
+.buttons button {
+  display: flex;
+  align-items: flex-start;
+  gap: 19px;
+  padding: 15px 0 15px 26px;
+  border: 1px solid var(--primary-color);
+  background-color: var(--secondary-color);
+  color: var(--text-color);
+  width: 100%;
+  cursor: pointer;
+  border-radius: 23px;
+  transition: background-color 0.3s ease;
+}
+
+.buttons button.active {
+  background-color: var(--primary-color);
+}
