+++ src/component/ProfileTrainer/styles.module.css
@@ -69,40 +69,6 @@
   border-radius: 50%;
 }
 
-.info {
-  position: absolute;
-  min-width: 200px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 30px;
-  bottom: -160%;
-}
-
-.info > div:first-child p:first-child {
-  font-size: 36px;
-  font-weight: 700;
-}
-.info > div:first-child p:nth-child(2) {
-  font-size: 32px;
-  font-weight: 500;
-}
-
-.info > div:first-child {
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-}
-
-.info > div:last-child {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  font-size: 24px;
-  font-weight: 600;
-}
-
 .main {
   display: flex;
   gap: 5.5rem;
@@ -135,29 +101,6 @@
   display: flex;
   flex-direction: column;
   gap: 30px;
-}
-
-.sideBarContainer button {
-  display: flex;
-  align-items: flex-start;
-  gap: 19px;
-  padding: 15px 0 15px 26px;
-  border: 1px solid var(--primary-color);
-  background-color: var(--secondary-color);
-  color: var(--text-color);
-  width: 100%;
-  cursor: pointer;
-  border-radius: 23px;
-  transition: background-color 0.3s ease;
-}
-
-.sideBarContainer button.active {
-  background-color: var(--primary-color);
-}
-
-.topBar,
-.personnelInfo {
-  display: none;
 }
 
 /* media Query */
@@ -219,45 +162,5 @@
   .userInfo > div > p {
     font-size: 1.5rem;
     font-weight: 600;
-  }
-
-  .topBar {
-    position: relative;
-    left: 50%;
-    transform: translateX(-50%);
-    display: flex;
-    text-align: center;
-    justify-content: center;
-    gap: 1rem;
-    padding: 1rem 2rem;
-    border-radius: 5px;
-    background-color: #f9f9f9;
-    width: fit-content;
-    transition: transform 3s ease, top 3s ease;
-  }
-
-  .topBar > div {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    min-width: 3rem;
-    height: 3rem;
-    border-radius: 12px;
-    background-color: var(--secondary-color);
-    border: 1px solid var(--primary-color);
-  }
-
-  .topBar div.active {
-    background-color: var(--primary-color);
-    color: white;
-    padding: 0 0.5rem;
-  }
-
-  .fixed {
-    position: sticky;
-    top: 0;
-    left: 50%;
-    transform: translate(-50%);
-    transition: transform 0.3s ease, top 0.3s ease;
   }
 }
